+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Split "display name" into "first name" and "last name"
 
 ## [v0.40.4] - 2020-09-05
+- Allow filtering by class
 
 - Added username to invitation link
 - Persist number of rows per page